requirements.txt
Pillow>=9.0.0
pyfiglet>=0.8.post1
colorama>=0.4.6